move `cdk.ts`, `stack.ts` and tests to conventional directory structure
