+++ .husky/pre-commit
@@ -7,7 +7,7 @@ yarn lint-staged
 
 if (git diff --name-only --cached | grep -e 'cdk' -e 'shared/graphql'); then
     yarn workspace cdk update-snapshot
-    git add cdk/test/__snapshots__/
+    git add cdk/lib/__snapshots__/
 fi
 
 if (git diff --name-only --cached | grep 'yarn.lock'); then
