+++ cdk/jest.config.js
@@ -1,8 +1,1 @@
-module.exports = {
-  roots: ["<rootDir>/test"],
-  testMatch: ["**/*.test.ts"],
-  preset: "ts-jest",
-  transform: {
-    "^.+\\.(j|t)sx?$": "ts-jest",
-  },
-};
+../jest.config.js
